+++ backend/windmill-worker/src/worker_flow.rs
@@ -29,7 +29,7 @@ use sqlx::types::Json;
 use sqlx::{FromRow, Postgres, Transaction};
 use tracing::instrument;
 use uuid::Uuid;
-use windmill_common::auth::JobPerms;
+use windmill_common::auth::get_job_perms;
 #[cfg(feature = "benchmark")]
 use windmill_common::bench::BenchmarkIter;
 use windmill_common::cache::{self, RawData};
@@ -626,6 +626,8 @@ pub async fn update_flow_status_after_job_completion_internal(
                              approvers: vec![],
                              failed_retries: vec![],
                              skipped: false,
+                             agent_actions: None,
+                             agent_actions_success: None,
                          }
                      } else {
                          success = false;
@@ -636,6 +638,8 @@ pub async fn update_flow_status_after_job_completion_internal(
                              flow_jobs_success: flow_jobs_success.clone(),
                              branch_chosen: None,
                              failed_retries: vec![],
+                             agent_actions: None,
+                             agent_actions_success: None,
                          }
                      };
                     let r = sqlx::query_scalar!(
@@ -806,6 +810,8 @@ pub async fn update_flow_status_after_job_completion_internal(
                             approvers: vec![],
                             failed_retries: old_status.retry.failed_jobs.clone(),
                             skipped: is_skipped,
+                            agent_actions: module_status.agent_actions(),
+                            agent_actions_success: module_status.agent_actions_success(),
                         }),
                     )
                 } else {
@@ -829,6 +835,8 @@ pub async fn update_flow_status_after_job_completion_internal(
                             flow_jobs_success,
                             branch_chosen,
                             failed_retries: old_status.retry.failed_jobs.clone(),
+                            agent_actions: module_status.agent_actions(),
+                            agent_actions_success: module_status.agent_actions_success(),
                         }),
                     )
                 }
@@ -2517,7 +2525,8 @@ async fn push_next_flow_job(
                 FlowModuleValue::Script { input_transforms, .. }
                 | FlowModuleValue::RawScript { input_transforms, .. }
                 | FlowModuleValue::FlowScript { input_transforms, .. }
-                | FlowModuleValue::Flow { input_transforms, .. },
+                | FlowModuleValue::Flow { input_transforms, .. }
+                | FlowModuleValue::AIAgent { input_transforms, .. },
             ) => {
                 let ctx = get_transform_context(&flow_job, &previous_id, &status)
                     .warn_after_seconds(3)
@@ -2605,6 +2614,8 @@ async fn push_next_flow_job(
                     approvers: vec![],
                     failed_retries: vec![],
                     skipped: false,
+                    agent_actions: None,
+                    agent_actions_success: None,
                 }),
                 flow_job.id
             )
@@ -2825,16 +2836,9 @@ async fn push_next_flow_job(
                 .flow_innermost_root_job
                 .or_else(|| Some(flow_job.id))
             {
-                sqlx::query_as!(
-                     JobPerms,
-                     "SELECT email, username, is_admin, is_operator, groups, folders FROM job_perms WHERE job_id = $1 AND workspace_id = $2",
-                     root_job,
-                     flow_job.workspace_id,
-                 )
-                 .fetch_optional(&mut *tx)
-                 .warn_after_seconds(3)
-                 .await?
-                 .map(|x| x.into())
+                get_job_perms(&mut *tx, root_job, &flow_job.workspace_id)
+                    .await?
+                    .map(|x| x.into())
             } else {
                 None
             }
@@ -2886,6 +2890,7 @@ async fn push_next_flow_job(
             Some(module.id.clone()),
             new_job_priority_override,
             job_perms.as_ref(),
+            false,
         )
         .warn_after_seconds(2)
         .await?;
@@ -3001,6 +3006,8 @@ async fn push_next_flow_job(
                 parallel: false,
                 while_loop,
                 progress: None,
+                agent_actions: None,
+                agent_actions_success: None,
             }
         }
         NextStatus::AllFlowJobs { iterator, branchall, .. } => FlowStatusModule::InProgress {
@@ -3014,6 +3021,8 @@ async fn push_next_flow_job(
             parallel: true,
             while_loop: false,
             progress: None,
+            agent_actions: None,
+            agent_actions_success: None,
         },
         NextStatus::NextBranchStep(NextBranch {
             mut flow_jobs,
@@ -3037,6 +3046,8 @@ async fn push_next_flow_job(
                 parallel: false,
                 while_loop: false,
                 progress: None,
+                agent_actions: None,
+                agent_actions_success: None,
             }
         }
 
@@ -3051,6 +3062,8 @@ async fn push_next_flow_job(
             parallel: false,
             while_loop: false,
             progress: None,
+            agent_actions: None,
+            agent_actions_success: None,
         },
         NextStatus::NextStep => {
             FlowStatusModule::WaitingForExecutor { id: status_module.id(), job: one_uuid? }
@@ -3253,12 +3266,12 @@ enum NextStatus {
 }
 
 #[derive(Clone)]
-struct JobPayloadWithTag {
-    payload: JobPayload,
-    tag: Option<String>,
-    delete_after_use: bool,
-    timeout: Option<i32>,
-    on_behalf_of: Option<OnBehalfOf>,
+pub struct JobPayloadWithTag {
+    pub payload: JobPayload,
+    pub tag: Option<String>,
+    pub delete_after_use: bool,
+    pub timeout: Option<i32>,
+    pub on_behalf_of: Option<OnBehalfOf>,
 }
 enum ContinuePayload {
     SingleJob(JobPayloadWithTag),
@@ -3320,7 +3333,7 @@ fn payload_from_modules<'a>(
     })
 }
 
-fn get_path(flow_job: &MiniPulledJob, status: &FlowStatus, module: &FlowModule) -> String {
+pub fn get_path(flow_job: &MiniPulledJob, status: &FlowStatus, module: &FlowModule) -> String {
     if status
         .preprocessor_module
         .as_ref()
@@ -3387,6 +3400,20 @@ async fn compute_next_flow_transform(
                 flow_to_payload(path, delete_after_use, &flow_job.workspace_id, db).await?;
             Ok(NextFlowTransform::Continue(
                 ContinuePayload::SingleJob(payload),
+                NextStatus::NextStep,
+            ))
+        }
+        FlowModuleValue::AIAgent { .. } => {
+            let path = get_path(flow_job, status, module);
+            let payload = JobPayload::AIAgent { path };
+            Ok(NextFlowTransform::Continue(
+                ContinuePayload::SingleJob(JobPayloadWithTag {
+                    payload,
+                    tag: None,
+                    delete_after_use,
+                    timeout: None,
+                    on_behalf_of: None,
+                }),
                 NextStatus::NextStep,
             ))
         }
@@ -4087,7 +4114,7 @@ async fn payload_from_simple_module(
     })
 }
 
-fn raw_script_to_payload(
+pub fn raw_script_to_payload(
     path: String,
     content: String,
     language: windmill_common::scripts::ScriptLang,
@@ -4137,7 +4164,7 @@ async fn flow_to_payload(
     Ok(JobPayloadWithTag { payload, tag, delete_after_use, timeout: None, on_behalf_of })
 }
 
-async fn script_to_payload(
+pub async fn script_to_payload(
     script_hash: Option<windmill_common::scripts::ScriptHash>,
     script_path: String,
     db: &sqlx::Pool<sqlx::Postgres>,