+++ frontend/src/lib/components/JobLoader.svelte
@@ -11,7 +11,8 @@
 		type FlowStatus,
 		type Preview,
 		type GetJobUpdatesResponse,
-		type WorkflowStatus
+		type WorkflowStatus,
+		type OpenFlow
 	} from '$lib/gen'
 	import { workspaceStore } from '$lib/stores'
 	import { onDestroy, tick, untrack } from 'svelte'
@@ -223,6 +224,25 @@
 					path: path ?? '',
 					requestBody: args,
 					skipPreprocessor: true
+				}),
+			callbacks
+		)
+	}
+
+	export async function runFlowPreview(
+		args: Record<string, any>,
+		flow: OpenFlow & { tag?: string },
+		callbacks?: Callbacks
+	): Promise<string> {
+		return abstractRun(
+			() =>
+				JobService.runFlowPreview({
+					workspace: $workspaceStore!,
+					requestBody: {
+						args,
+						value: flow.value,
+						tag: flow.tag
+					}
 				}),
 			callbacks
 		)
@@ -577,6 +597,7 @@
 						noLogs: noLogs,
 						noCode
 					})
+
 					callbacks?.change?.(job)
 				}