+++ frontend/src/lib/components/ModuleTest.svelte
@@ -1,11 +1,12 @@
 <script lang="ts">
-	import { ScriptService, type FlowModule, type Job } from '$lib/gen'
+	import { ScriptService, type FlowModule, type JavascriptTransform, type Job } from '$lib/gen'
 	import { workspaceStore } from '$lib/stores'
 	import { getScriptByPath } from '$lib/scripts'
 	import { getContext } from 'svelte'
 	import type { FlowEditorContext } from './flows/types'
 	import JobLoader, { type Callbacks } from './JobLoader.svelte'
 	import { getStepHistoryLoaderContext } from './stepHistoryLoader.svelte'
+	import { loadSchemaFromModule } from './flows/flowInfers'
 
 	interface Props {
 		mod: FlowModule
@@ -38,7 +39,7 @@
 
 	export function loadArgsAndRunTest() {
 		stepsInputArgs?.updateStepArgs(mod.id, flowStateStore.val, flowStore?.val, previewArgs?.val)
-		runTest(stepsInputArgs.getStepArgs(mod.id)?.value)
+		runTest(stepsInputArgs.getStepArgs(mod.id))
 	}
 
 	export async function runTest(args: any) {
@@ -86,6 +87,39 @@
 			)
 		} else if (val.type == 'flow') {
 			await jobLoader?.runFlowByPath(val.path, args, callbacks)
+		} else if (val.type == 'aiagent') {
+			const { schema } = await loadSchemaFromModule(mod)
+
+			const inputTransforms: { [key: string]: JavascriptTransform } = Object.fromEntries(
+				Object.keys(args).map((key) => [
+					key,
+					{
+						expr: `flow_input.${key}`,
+						type: 'javascript'
+					}
+				])
+			)
+
+			await jobLoader?.runFlowPreview(
+				args,
+				{
+					value: {
+						modules: [
+							{
+								...mod,
+								value: {
+									type: 'aiagent',
+									tools: mod.value.type == 'aiagent' ? mod.value.tools : [],
+									input_transforms: inputTransforms
+								}
+							}
+						]
+					},
+					summary: '',
+					schema
+				},
+				callbacks
+			)
 		} else {
 			throw Error('Not supported module type')
 		}
@@ -148,7 +182,10 @@
 			}
 		}
 	}
-	bind:job={modulesTestStates.states[mod.id].testJob}
+	bind:job={
+		() => modulesTestStates.states[mod.id]?.testJob,
+		(v) => modulesTestStates.states[mod.id] && (modulesTestStates.states[mod.id].testJob = v)
+	}
 	loadPlaceholderJobOnStart={{
 		type: 'QueuedJob',
 		id: '',